Cleanup of packages, simplify AppContext, update initial user selection
I'd added unnecessary "fancy" to the user database functions which didn't help readability and just made things a little too complicated.
The initial user selection screen is now updated to look like the wireframes we've discussed in meetings and drawn on the board. The interaction pattern was also simplified to allow one-touch action rather than selection and acknowledgement. We can provide a small "back" button on the eventual welcome screen that goes back to the user selection to prevent accidents.
There were also some unnecessary buttons on the Dashboard view
Updated the spacing/positioning on the CompactRadio component for a little nicer layout