Simplified the setting and sending of the ETA info
Stumbled across a method, that I should have realized before now, which allows the ServiceLayer to just automatically respond to changes in the app state. App state comes in as props via AppContextProvider which we can then just compare for differences, acting on anything that's different (duh). This is what can fire the internal methods which communicate with the other app.
Also fancied-up the duration count down display.