Refactor configs, simplify URL, add configLoader
Refactored the config loading into a source file in the App folder so the `config.js` file can be simpler (no real executing code, just an object). Applied this to console and copilot source files.
Simplified the URL so only the query string keys you provide will be re-printed back to the URL, making for "prettier" URLs and not exposing potentially private keys to the URL and history. It's a minor bonus but it's a little nicer to work with. Also, only query string arguments will get saved to localStorage, which allows the lowest level config.js file to still exert some influence on the app(s) if it's changed and the URL arg for that key was never manually set.
Also a little fix so the `/?` isn't added if there are no arguments to add to the URL.